% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbp_variability.R
\name{add_variability}
\alias{add_variability}
\title{Add Variability Features to Session Data}
\usage{
add_variability(sessions, variability)
}
\arguments{
\item{sessions}{Output of [summarize_sessions()].}

\item{variability}{Output of [session_variability()].}
}
\value{
`sessions` with the feature columns added and listed in attribute `"parameters"`.
}
\description{
Joins the output of [session_variability()] to [summarize_sessions()] output so
the features can be analyzed like respiratory parameters.
}
