% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_by_parameter.R
\name{auc_by_parameter}
\alias{auc_by_parameter}
\title{Calculate Area Under the Curve (AUC) for Each Parameter by Group}
\usage{
auc_by_parameter(groups_merged_named, group_to_normalize = NULL)
}
\arguments{
\item{groups_merged_named}{A list of data frames, where each element represents a group, and each data frame
contains a `Time` column and numeric variables (parameters) to compute AUC for.}

\item{group_to_normalize}{A character string specifying the group whose AUC values will be used for normalization.
If `NULL`, no normalization is performed. Default is `NULL`.}
}
\value{
A data frame containing the following columns:
\itemize{
  \item \strong{Group}: The name of the group.
  \item \strong{Parameter}: The name of the parameter.
  \item \strong{AUC}: The computed Area Under the Curve for the parameter in the group.
  \item \strong{AUC_sd}: The standard deviation of the AUC values.
  \item \strong{AUC_normalized} (optional): The normalized AUC value, if `group_to_normalize` is provided.
}
}
\description{
This function calculates the Area Under the Curve (AUC) for each parameter over time for different groups.
The AUC is computed using the trapezoidal rule, and the standard deviation of the AUC is also calculated.
Optionally, the AUC values can be normalized to a specified group.
}
\details{
This function computes the AUC for each parameter by applying the trapezoidal method, which approximates the area under the
curve defined by each parameter's time series data. The function also computes the standard deviation (SD) of the AUC values.
If `group_to_normalize` is provided, the AUC values of all groups are normalized to the AUC of the specified group for each
parameter. The function returns a data frame containing the AUC values, AUC standard deviations, and optionally the
normalized AUC values.
}
\examples{
# Assuming groups_merged_named is a list of data frames, each containing Time and parameters
auc_results <- auc_by_parameter(groups_merged_named)

# Calculate AUC and normalize by the specified group
auc_results_normalized <- auc_by_parameter(groups_merged_named, group_to_normalize = "Group A")

}
