% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_wbp_heatmap.R
\name{auc_wbp_heatmap}
\alias{auc_wbp_heatmap}
\title{Generate and Save a Heatmap of Normalized AUC Values by Group and Parameter}
\usage{
auc_wbp_heatmap(
  auc_results_normalized,
  group_to_remove = NULL,
  remove_parameters = NULL,
  main_title = "Normalized AUC by Group and Parameter",
  save_heatmap = FALSE
)
}
\arguments{
\item{auc_results_normalized}{A data frame containing the normalized AUC results. The data frame should include the columns `Group`, `Parameter`, and `AUC_normalized`.}

\item{group_to_remove}{A character string specifying a group to be removed from the heatmap data. Default is `NULL`. If specified, this group will be excluded from the heatmap.}

\item{remove_parameters}{A vector of character strings specifying the parameters to be removed from the heatmap data. Default is `NULL`. If specified, these parameters will be excluded from the heatmap.}

\item{main_title}{A character string specifying the title of the heatmap. Default is "Normalized AUC by Group and Parameter".}

\item{save_heatmap}{A logical value indicating whether to save the heatmap as a PNG file. Default is `FALSE`. If `TRUE`, the heatmap will be saved in the current working directory.}
}
\value{
A heatmap plot displayed in the RStudio plot window. If `save_heatmap` is `TRUE`, the plot is saved as a PNG file.
}
\description{
This function generates a heatmap of normalized Area Under the Curve (AUC) values for each parameter, grouped by experimental conditions (e.g., different groups). The function allows for optional removal of specific groups and parameters, and it can save the generated heatmap as a PNG file.
}
\details{
This function creates a heatmap where the rows represent different parameters and the columns represent different groups. The values in the heatmap are the normalized AUC values. The data is clustered both by rows and columns using hierarchical clustering, with Euclidean distance and complete linkage. The heatmap is generated using the `pheatmap` package.

If `save_heatmap` is `TRUE`, the heatmap is saved as a PNG file with the title specified by `main_title`. The file will be saved in the current working directory with the title formatted as "main_title_heatmap.png".
}
\examples{
# Generate and display a heatmap without saving
auc_wbp_heatmap(auc_results_normalized)

# Generate and display a heatmap, removing specific group and parameters, and saving the heatmap as a PNG file
auc_wbp_heatmap(auc_results_normalized, group_to_remove = "Group A",
                remove_parameters = c("Comp", "EEP"), main_title = "Test", save_heatmap = TRUE)

}
