% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_group.R
\name{define_group}
\alias{define_group}
\title{Define Groups for Data}
\usage{
define_group(num_groups = NULL, items_per_group = NULL, custom_sizes = NULL)
}
\arguments{
\item{num_groups}{Integer specifying the number of groups to create. Required if `custom_sizes` is not provided.}

\item{items_per_group}{Integer specifying the number of items in each group. Required if `custom_sizes` is not provided.}

\item{custom_sizes}{Numeric vector specifying the custom sizes for each group. If provided, `num_groups` and `items_per_group` should not be used.}
}
\value{
A list where each element represents a group and contains the indices of the items in that group.
}
\description{
This function allows you to define groups of items based on either a fixed number of groups and items per group, or custom group sizes.
It returns a list where each element contains the indices for a group.
}
\details{
- If `custom_sizes` is provided, the function generates groups with sizes specified by the vector.
- If `num_groups` and `items_per_group` are provided, the function generates groups with the specified number of groups and items per group.
- If neither `custom_sizes` nor both `num_groups` and `items_per_group` are provided, an error is raised.
}
\examples{
# Define 4 groups with 4 items each
group_identities <- define_group(num_groups = 4, items_per_group = 4)

# Define 5 groups with custom sizes (3, 1, 4, 4, 4)
group_identities <- define_group(num_groups = 5, custom_sizes = c(3, 1, 4, 4, 4))

}
