% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbp_design.R
\name{design_dpi}
\alias{design_dpi}
\title{Days Post-Infection of Each Session}
\usage{
design_dpi(sessions, design)
}
\arguments{
\item{sessions}{Output of [summarize_sessions()].}

\item{design}{Output of [read_design()] with `infection_session` set.}
}
\value{
Named numeric vector: mean days post-infection of each timepoint.
}
\description{
Days Post-Infection of Each Session
}
