% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbp_library.R
\name{library_init}
\alias{library_init}
\title{Create or Open a Study Library}
\usage{
library_init(lib_dir)
}
\arguments{
\item{lib_dir}{Folder for the library (created if needed).}
}
\value{
`lib_dir`, invisibly.
}
\description{
Create or Open a Study Library
}
