% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbp_library.R
\name{library_list}
\alias{library_list}
\title{List the Studies in a Library}
\usage{
library_list(lib_dir)
}
\arguments{
\item{lib_dir}{Library folder.}
}
\value{
A data frame with one row per study (empty if none).
}
\description{
List the Studies in a Library
}
