% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbp_library.R
\name{library_load_model}
\alias{library_load_model}
\title{Load a Saved Model}
\usage{
library_load_model(lib_dir, model_id)
}
\arguments{
\item{lib_dir}{Library folder.}

\item{model_id}{Model id from [library_models()].}
}
\value{
The saved `plethr_ml` object.
}
\description{
Load a Saved Model
}
