% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbp_library.R
\name{library_models}
\alias{library_models}
\title{List Saved Models}
\usage{
library_models(lib_dir)
}
\arguments{
\item{lib_dir}{Library folder.}
}
\value{
A data frame with one row per saved model (empty if none).
}
\description{
List Saved Models
}
