% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbp_library.R
\name{library_remove}
\alias{library_remove}
\title{Remove a Study from the Library}
\usage{
library_remove(lib_dir, study_id)
}
\arguments{
\item{lib_dir}{Library folder.}

\item{study_id}{Study to remove.}
}
\value{
The updated study index, invisibly.
}
\description{
Remove a Study from the Library
}
