% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbp_library.R
\name{library_save_model}
\alias{library_save_model}
\title{Save a Model in the Library}
\usage{
library_save_model(lib_dir, object, name, notes = "")
}
\arguments{
\item{lib_dir}{Library folder.}

\item{object}{Output of [ml_fit()].}

\item{name}{Short name for the model.}

\item{notes}{Free text.}
}
\value{
The model id, invisibly.
}
\description{
Save a Model in the Library
}
