% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_group_avg.R
\name{merge_group_avg}
\alias{merge_group_avg}
\title{Merge and Average Data by Group}
\usage{
merge_group_avg(df_list_clean, group_identities, group_names = NULL)
}
\arguments{
\item{df_list_clean}{A list of cleaned data frames. Each data frame should contain a `Time` column and other numeric columns.}

\item{group_identities}{A list of numeric vectors specifying the indices of the data frames to be merged for each group.}

\item{group_names}{A character vector of names for each group. If provided, this will be used to name the resulting list of averaged groups. Default is `NULL`.}
}
\value{
A list of data frames, where each element represents a group. Each data frame contains the average values of
the columns for each `Time` value within the group.
}
\description{
This function merges multiple data frames into groups based on specified group identities, and then averages the data
within each group by the `Time` variable. It is useful for summarizing data across different groups in a time series context.
}
\details{
The function will take the data frames specified by `sheets_into_list`, merge them by group, and then calculate the mean for each
numeric column, grouped by `Time`. group_identities must be predefined by the define_group() function to properly summarize by group.
If `group_names` is provided, the resulting list of averaged data frames will be named according to the given `group_names` as previously defined by set_group_names().
}
\examples{
# Assuming df_list_summarized is a list of cleaned data frames and group_identities is a list of group indices
groups_merged <- merge_group_avg(cp_df_list_summarized, group_identities)

# If you want to name the groups, provide the group names
groups_merged_named <- merge_group_avg(df_list_summarized, group_identities, group_names)

}
