% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbp_library.R
\name{ml_study_metrics}
\alias{ml_study_metrics}
\title{Performance of a Cross-Study Model in Each Study}
\usage{
ml_study_metrics(object)
}
\arguments{
\item{object}{Output of [ml_fit()] on library data.}
}
\value{
A data frame: model, study, session-level and animal-level ROC AUC and accuracy.
}
\description{
Performance of a Cross-Study Model in Each Study
}
