% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plethR_packages.R
\name{plethR_packages}
\alias{plethR_packages}
\title{Install and Load Required Packages}
\usage{
plethR_packages()
}
\description{
This function installs and loads all the necessary packages for the package development and usage.
}
