% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_auc_by_parameter.R
\name{plot_auc_by_parameter}
\alias{plot_auc_by_parameter}
\title{Generate and Save Bar Plots of AUC by Parameter}
\usage{
plot_auc_by_parameter(auc_results, save_plots = FALSE, y_scale = FALSE)
}
\arguments{
\item{auc_results}{A data frame containing the AUC results, including columns for `Group`, `Parameter`, `AUC`, and `AUC_sd`.}

\item{save_plots}{A logical value indicating whether to save the plots as PNG files. Default is `FALSE`.
If `TRUE`, the plots will be saved in a folder named "AUC_Results_parameter_bar_plots".}

\item{y_scale}{A logical value indicating whether to automatically scale the y-axis based on AUC values and standard deviations.
Default is `FALSE`. If `TRUE`, the y-axis will be scaled to accommodate the full range of AUC values and their standard deviations.}
}
\value{
A list of `ggplot` objects, each corresponding to a plot of AUC for a parameter.
}
\description{
This function generates bar plots of Area Under the Curve (AUC) values for each parameter in the provided AUC results,
with optional error bars and normalization. The plots can be saved to disk if specified.
}
\details{
This function generates a separate bar plot for each unique parameter in the `auc_results` data frame. Each plot represents
the AUC values for different groups, with error bars representing the standard deviation of the AUC. The color scheme used
in the plots is colorblind-friendly, derived from the `RColorBrewer` package.
The plots are generated using `ggplot2`. If `save_plots` is set to `TRUE`, the plots are saved as PNG files in a directory
named "AUC_Results_parameter_bar_plots".
}
\examples{
# Assuming `auc_results_normalized` is a data frame with AUC results
# Generate and display plots without saving
plot_list <- plot_auc_by_parameter(auc_results_normalized)

# Generate, scale, and save plots
plot_list_saved <- plot_auc_by_parameter(auc_results_normalized, save_plots = TRUE, y_scale = TRUE)

}
