% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_group.R
\name{plot_by_group}
\alias{plot_by_group}
\title{Plot Time Series Data by Group}
\usage{
plot_by_group(groups_merged_named, rolling_average = NULL, save_plots = FALSE)
}
\arguments{
\item{groups_merged_named}{A list of data frames, where each element represents a group, containing a `Time` column and
numeric variables that will be plotted.}

\item{rolling_average}{An optional integer specifying the window size for the rolling average. If `NULL`, no rolling average
is applied. Default is `NULL`.}

\item{save_plots}{A logical indicating whether to save the generated plots as PNG files. If `TRUE`, plots will be saved to
a folder named after the base name of the input file. Default is `FALSE`.}
}
\value{
A list of ggplot objects, each corresponding to a plot for one group.
}
\description{
This function generates line plots for each group over time, with the option to apply a rolling average to smooth the data.
The function also supports saving the resulting plots to a specified folder.
}
\details{
This function takes the provided list of data frames (`groups_merged_named`) from merge_group_avg(), where each data frame corresponds to a group,
and generates line plots for each unique variable in the data, grouped by time. The plots are faceted by variable, and the
function supports the application of a rolling average for smoothing. If `save_plots` is `TRUE`, the plots are saved to a
folder named based on the input file, without the `.xlsx` extension.
}
\examples{
# Assuming `groups_merged_named` is a list of data frames, each containing a Time column and variables
plot_list <- plot_by_group(groups_merged_named, rolling_average = 5, save_plots = FALSE)

# If you want to save the plots to files
plot_list <- plot_by_group(groups_merged_named, rolling_average = 5, save_plots = TRUE)

}
