% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_parameter.R
\name{plot_by_parameter}
\alias{plot_by_parameter}
\title{Plot Time Series Data by Parameter}
\usage{
plot_by_parameter(
  groups_merged_named,
  rolling_average = NULL,
  save_plots = FALSE
)
}
\arguments{
\item{groups_merged_named}{A list of data frames, where each element represents a group and contains a `Time` column and
numeric parameters to be plotted.}

\item{rolling_average}{An optional integer specifying the window size for the rolling average. If `NULL`, no rolling average
is applied. Default is `NULL`.}

\item{save_plots}{A logical indicating whether to save the generated plots as PNG files. If `TRUE`, plots will be saved to
a folder named after the base name of the input file. Default is `FALSE`.}
}
\value{
A list of ggplot objects, each corresponding to a plot for one parameter.
}
\description{
This function generates line plots of parameters over time for different groups. It supports the option to apply a rolling
average to smooth the data and can save the resulting plots to a specified folder.
}
\details{
This function takes the provided list of data frames (`groups_merged_named`) from merge_group_avg() and generates line plots for each unique parameter
in the data, grouped by time. The plots can be customized to show a rolling average of the values for each parameter. If
`save_plots` is `TRUE`, the function will save each plot as a PNG file in a folder named after the base name of the provided
`excel_file` (if applicable).
}
\examples{
# Assuming `groups_merged_named` is a list of data frames, each containing a Time column and parameters
plot_list <- plot_by_parameter(groups_merged_named, rolling_average = 2, save_plots = FALSE)

# If you want to save the plots to files
plot_list <- plot_by_parameter(groups_merged_named, rolling_average = 5, save_plots = TRUE)

}
