% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbp_design.R
\name{plot_cfu_correlation}
\alias{plot_cfu_correlation}
\title{Plot Breathing against Bacterial Burden}
\usage{
plot_cfu_correlation(correlation, parameter, colors = NULL)
}
\arguments{
\item{correlation}{Output of [cfu_correlation()].}

\item{parameter}{Parameter to plot.}

\item{colors}{Named group colors.}
}
\value{
A ggplot object.
}
\description{
Plot Breathing against Bacterial Burden
}
