% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbp_ml.R
\name{plot_ml_importance}
\alias{plot_ml_importance}
\title{Plot Variable Importance}
\usage{
plot_ml_importance(object, model, n = 15)
}
\arguments{
\item{object}{Output of [ml_fit()].}

\item{model}{Model code (`"rf"`, `"bt"`, `"en"`, `"log"` or `"lm"`).}

\item{n}{Number of features to show.}
}
\value{
A ggplot object.
}
\description{
Plot Variable Importance
}
