% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbp_library.R
\name{plot_model_history}
\alias{plot_model_history}
\title{Plot Model History}
\usage{
plot_model_history(models)
}
\arguments{
\item{models}{Output of [library_models()].}
}
\value{
A ggplot object.
}
\description{
Held-out performance of saved models over time, to see whether models improve
as studies are added.
}
