% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbp_mixed.R
\name{plot_power_curve}
\alias{plot_power_curve}
\title{Plot Power against Group Size}
\usage{
plot_power_curve(plan, parameter, colors = NULL, max_n = 40)
}
\arguments{
\item{plan}{Output of [sample_size_plan()].}

\item{parameter}{Parameter to show (one curve per group compared with the reference).}

\item{colors}{Named group colors.}

\item{max_n}{Largest group size on the x axis.}
}
\value{
A ggplot object.
}
\description{
Plot Power against Group Size
}
