% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbp_treatment.R
\name{plot_treatment_parameters}
\alias{plot_treatment_parameters}
\title{Plot Treatment Effect per Parameter}
\usage{
plot_treatment_parameters(efficacy)
}
\arguments{
\item{efficacy}{Output of [treatment_efficacy()].}
}
\value{
A ggplot object.
}
\description{
Rescue (share of the disease effect removed by treatment) for each parameter.
}
