% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_into_list.R
\name{sheets_into_list}
\alias{sheets_into_list}
\title{Load and Process Excel Sheets into a List of Data Frames}
\usage{
sheets_into_list(
  excel_file,
  remove_apnea = FALSE,
  clean_time = FALSE,
  date_average = FALSE,
  remove_na = FALSE
)
}
\arguments{
\item{excel_file}{A character string specifying the path to the Excel file.}

\item{remove_apnea}{Logical, if `TRUE`, removes sheets that contain "Apnea" in their names. Default is `FALSE`.}

\item{clean_time}{Logical, if `TRUE`, formats the "Time" column as a date in the format `YYYY-MM-DD`. Default is `FALSE`.}

\item{date_average}{Logical, if `TRUE`, averages all columns by date in the "Time" column. Requires `clean_time = TRUE` to ensure the "Time" column is formatted correctly. Default is `FALSE`.}

\item{remove_na}{Logical, if `TRUE`, removes rows with missing values in each sheet. Default is `FALSE`.}
}
\value{
A list of data frames, with each data frame representing one processed sheet from the Excel file.
}
\description{
This function loads all sheets from an Excel file into a list of data frames,
with options to apply several preprocessing steps, including removing specific sheets,
selecting relevant columns, cleaning date formats, removing rows with missing values,
and averaging data by date.
}
\details{
This function loads all sheets in the specified Excel file into a list of data frames.
Several options allow for data cleaning and preprocessing:

- `remove_apnea`: If `TRUE`, any sheet with "Apnea" in its name is excluded from the list.
- `clean_time`: If `TRUE`, the "Time" column (if present) is converted to a date format.
- `remove_na`: If `TRUE`, rows containing `NA` values are removed from each sheet.
- `date_average`: If `TRUE`, the data is averaged by date, with each unique date in the "Time" column representing a summary row.
}
\examples{
\dontrun{
excel_file <- "example.xlsx"

# Load all sheets from an Excel file and remove any sheets containing "Apnea" in their name
df_list <- sheets_into_list(excel_file, remove_apnea = TRUE)

# Load sheets and clean the "Time" column, remove rows with NA, and average data by date
df_list <- sheets_into_list(excel_file, clean_time = TRUE, date_average = TRUE,
                            remove_na = TRUE)
}

}
