% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbp_plots.R
\name{theme_plethr}
\alias{theme_plethr}
\title{Publication Theme for plethR Figures}
\usage{
theme_plethr(base_size = 13)
}
\arguments{
\item{base_size}{Base font size in points.}
}
\value{
A ggplot2 theme.
}
\description{
Publication Theme for plethR Figures
}
