% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbp_variability.R
\name{variability_feature_types}
\alias{variability_feature_types}
\title{Breathing Variability Feature Types}
\usage{
variability_feature_types()
}
\value{
A named character vector: feature codes and their descriptions.
}
\description{
Breathing Variability Feature Types
}
\examples{
variability_feature_types()
}
