% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_define_group.R
\name{custom_define_group}
\alias{custom_define_group}
\title{Assign Sheet Numbers to Groups Interactively with Optional String Mapping}
\usage{
custom_define_group(group_names, use_string_value = FALSE)
}
\arguments{
\item{group_names}{Character vector specifying the names of the groups. Each name will be
prompted to collect the corresponding sheet numbers.}

\item{use_string_value}{Logical. If `TRUE`, sheet numbers will be mapped to string identifiers
(e.g., "1.WBPth"). Default is `FALSE`.}
}
\value{
A named list where each element corresponds to a group and contains the assigned sheet numbers
or string identifiers, depending on the value of `use_string_value`.
}
\description{
This function allows users to assign sheet numbers to predefined groups interactively.
It returns a named list where each element represents a group and contains the indices
of the sheets assigned to that group. An optional feature allows for mapping sheet numbers
to string identifiers (e.g., "1.WBPth") if `use_string_value` is set to `TRUE`.
}
\details{
- The function prompts the user to input sheet numbers for each group interactively. Sheet numbers
should be entered as a comma-separated list (e.g., "1,3,5").
- Duplicate sheet assignments across groups are not allowed and will result in an error.
- The sheet numbers are assumed to correspond to the order in the input data (e.g., Sheet1 = 1, Sheet2 = 2, etc.).
- User should NOT insert spaces between comma-separated values.
- If `use_string_value` is `TRUE`, the sheet numbers will be appended with ".WBPth" to map them to string identifiers.
}
\examples{
# Requires user predefined group names using set_group_names function.
test_names <- set_group_names(c("Group A", "Group B", "Group C", "Group D"))

# Interactively assign sheet numbers to these groups with string identifiers
group_mapping <- custom_define_group(group_names = test_names, use_string_value = TRUE)

}
