% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbp_ml.R
\name{ml_model_names}
\alias{ml_model_names}
\title{Machine Learning Model Names}
\usage{
ml_model_names()
}
\value{
A named character vector: short codes (names) and full model names.
}
\description{
Machine Learning Model Names
}
\examples{
ml_model_names()
}
