% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbp_ml.R
\name{plot_ml_confusion}
\alias{plot_ml_confusion}
\title{Plot a Confusion Matrix}
\usage{
plot_ml_confusion(object, model, level = c("session", "animal"))
}
\arguments{
\item{object}{Output of [ml_fit()].}

\item{model}{Model code.}

\item{level}{`"session"` or `"animal"`.}
}
\value{
A ggplot object.
}
\description{
Plot a Confusion Matrix
}
