% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbp_ml.R
\name{plot_ml_roc}
\alias{plot_ml_roc}
\title{Plot ROC Curves}
\usage{
plot_ml_roc(object, level = c("session", "animal"), models = NULL)
}
\arguments{
\item{object}{Output of [ml_fit()].}

\item{level}{`"session"` or `"animal"`.}

\item{models}{Models to show. Defaults to all.}
}
\value{
A ggplot object.
}
\description{
Plot ROC Curves
}
