% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_group_names.R
\name{set_group_names}
\alias{set_group_names}
\title{Set Group Names}
\usage{
set_group_names(names_vector)
}
\arguments{
\item{names_vector}{A character vector containing group names.}
}
\value{
The same character vector passed as input, containing the group names.
}
\description{
This function takes a character vector and returns it as-is, with an error check to ensure the input is a character vector.
}
\examples{
# Example usage:
group_names <- set_group_names(c("Uninfected Wildtype", "Uninfected Benac", "Infected Benac", "Infected Wildtype"))
print(group_names)

}
