% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbp_pca_plot.R
\name{wbp_pca_plot}
\alias{wbp_pca_plot}
\title{PCA Plot with User Defined K-means Clustering and Optional Saving}
\usage{
wbp_pca_plot(
  df_list,
  num_clusters = 4,
  text_size = 5,
  save_plots = FALSE,
  random_seed = 2
)
}
\arguments{
\item{df_list}{Input the object output from 'sheets_into_list()' step. Each data frame should contain numeric columns only.}

\item{num_clusters}{Integer specifying the number of clusters for k-means clustering. Default is 4.}

\item{text_size}{Numeric value for the text size of labels on the plot. Default is 5.}

\item{save_plots}{Logical. If `TRUE`, saves the plot as "pca_plot.png" in the current directory. Default is `FALSE`.}

\item{random_seed}{Integer specifiying the random seed for reproducibility. Default is 1. Required because K-means algorithm involves random initialization step for cluster centroids, which can lead to different results each run.}
}
\value{
A ggplot object showing the PCA plot with clusters and ellipses.
}
\description{
This function performs Principal Component Analysis (PCA) on a list of data frames, applies k-means clustering,
and creates a 2D PCA plot of the first two components with clusters highlighted by ellipses.
It also allows for optional saving of the plot as a PNG file.
}
\examples{
# Assuming object is a list of numeric data frames
wbp_pca_plot(df_list_summarized, num_clusters = 3, text_size = 4, save_plots = TRUE)

}
